requirements.txt
langchain
python-dotenv
streamlit
unstructured
faiss-cpu
libmagic
python-magic
langchain_groq
sentence-transformers

